refactor: move alignment param calculation into a separate file
To make it testable
